联盟赛上场版
自瞄有点bug不好复现，自瞄改成连续值，当丢目标或者没识别到目标时有概率云台会转会初始状态

+++ VT13/NEW_VISION/GIMBAL/User/Algorithm/inc/QuaternionEKF.h
@@ -73,7 +73,7 @@ typedef struct
     float ChiSquareTestThreshold; // 卡方检验阈值
     float lambda;                 // 渐消因子
 
-    int16_t YawRoundCount;
+    float YawRoundCount;
 
     float YawAngleLast;
 } QEKF_INS_t;
